refactor(Model): Model 폴더 삭제 및 클래스 경로 수정
요구 사항 중, Lotto.js의 패키지를 변경을 제한했으므로, Models 폴더 내 모든 파일을 src로 이동
